do not insert reference path any longer, do not consider kmers that do not occur on any of the paths anymore, as they will not change the probabilities (always cn=0). Changed Variant implementation to keep track of the reference allele, also in cases where no path covers it.

+++ src/copynumber.cpp
@@ -8,11 +8,11 @@ CopyNumber::CopyNumber()
 :probabilities({1.0,0.0,0.0})
 {}
 
-CopyNumber::CopyNumber(double cn_0, double cn_1, double cn_2)
+CopyNumber::CopyNumber(long double cn_0, long double cn_1, long double cn_2)
 	:probabilities({cn_0, cn_1, cn_2})
 {}
 
-double CopyNumber::get_probability_of(int cn){
+long double CopyNumber::get_probability_of(int cn){
 	if( (cn < 0) || (cn > 2) ){
 		ostringstream oss;
 		oss << "CopyNumber::get_probability_of: Invalid copy number: " << cn;
